rename folders to better readability
