+++ apps/meteor/ee/server/local-services/federation/application/UserService.ts
@@ -1,6 +1,6 @@
 import type { IFederationBridgeEE } from '../domain/IFederationBridge';
 import type { RocketChatUserAdapterEE } from '../infrastructure/rocket-chat/adapters/User';
-import { AbstractFederationApplicationServiceEE } from './AbstractFederationService';
+import { AbstractFederationApplicationServiceEE } from './AbstractFederationApplicationServiceEE';
 import type { RocketChatSettingsAdapter } from '../../../../../server/services/federation/infrastructure/rocket-chat/adapters/Settings';
 import type { RocketChatFileAdapter } from '../../../../../server/services/federation/infrastructure/rocket-chat/adapters/File';
