+++ apps/meteor/ee/server/local-services/federation/application/room/sender/DirectMessageRoomServiceSender.ts
@@ -8,7 +8,7 @@ import type {
 	FederationOnDirectMessageRoomCreationDto,
 	FederationRoomInviteUserDto,
 } from './input/RoomSenderDto';
-import { AbstractFederationApplicationServiceEE } from '../../AbstractFederationService';
+import { AbstractFederationApplicationServiceEE } from '../../AbstractFederationApplicationServiceEE';
 import type { RocketChatFileAdapter } from '../../../../../../../server/services/federation/infrastructure/rocket-chat/adapters/File';
 import type { RocketChatSettingsAdapter } from '../../../../../../../server/services/federation/infrastructure/rocket-chat/adapters/Settings';
