+++ apps/meteor/server/services/federation/application/room/message/receiver/MessageServiceReceiver.ts
@@ -6,7 +6,7 @@ import type { RocketChatMessageAdapter } from '../../../../infrastructure/rocket
 import type { RocketChatRoomAdapter } from '../../../../infrastructure/rocket-chat/adapters/Room';
 import type { RocketChatSettingsAdapter } from '../../../../infrastructure/rocket-chat/adapters/Settings';
 import type { RocketChatUserAdapter } from '../../../../infrastructure/rocket-chat/adapters/User';
-import { AbstractFederationApplicationService } from '../../../AbstractFederationService';
+import { AbstractFederationApplicationService } from '../../../AbstractFederationApplicationServicetionService';
 import type { FederationMessageReactionEventDto } from '../../input/MessageReceiverDto';
 
 export class FederationMessageServiceReceiver extends AbstractFederationApplicationService {
